requirements.txt
Flask==3.0.0
python-dotenv==1.0.0
groq==0.4.1
requests==2.32.3
gunicorn==21.2.0
